Made sure API is not broken.
In the previous commit, the API was broken by adding a new parameter
"mimetype" to the function pico_http_respond. This isn't an optimal
solution. Instead, the function with the extra parameter was renamed to
pico_http_respond_mimetype. The implementation of the existing
pico_http_respond function was improved by adding a guessing algorithm
for the mimetype